requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
python-crontab
passlib
paramiko
psutil

